move templates-core to packages
