+++ packages/antd-tables/index.js
@@ -1,7 +1,7 @@
 const {
   TemplatePackage,
   ChartRendererSnippet,
-} = require('../../templates-core');
+} = require('@cubejs-templates/core');
 
 class AntdTablesTemplate extends TemplatePackage {}
