+++ packages/static-chart/index.js
@@ -1,10 +1,10 @@
-const { TemplatePackage, ChartSnippet } = require('../../templates-core');
+const { TemplatePackage, ChartSnippet } = require('@cubejs-templates/core');
 
 class StaticChartTemplate extends TemplatePackage {}
 
 module.exports = (context) => {
   const { chartCode } = context.playgroundContext;
-  
+
   if (!chartCode) {
     throw new Error(`playgroundContext misses required chartCode`);
   }
