+++ packages/templates-core/.gitignore
