+++ packages/templates-core/index.js
