+++ packages/templates-core/package.json
