+++ packages/templates-core/src/AppSnippet.js
