+++ packages/templates-core/src/ChartRendererSnippet.js
