+++ packages/templates-core/src/ChartSnippet.js
