+++ packages/templates-core/src/IndexSnippet.js
