+++ packages/templates-core/src/StaticChartTemplate.js
