+++ packages/templates-core/src/TargetSource.js
