+++ packages/templates-core/src/TemplatePackage.js
